requirements.txt
pandas
streamlit==1.36.0
tiktoken==0.7.0
transformers==4.44.2
torch>=2.2.0
seaborn
matplotlib
statsmodels
scipy
google-generativeai
tokenizers
